change bugfixWords metrics to ratio

+++ src/main/kotlin/io/github/durun/nitron/inout/model/metrics/metricsTables.kt
@@ -40,6 +40,6 @@ object GlobalPatternsTable : Table("globalPatterns") {
     val dChars: Column<Int> = integer("dChars")
     val dTokens: Column<Int> = integer("dTokens")
     val authors: Column<Int> = integer("authors")
-    val bugfixWords: Column<Int> = integer("bugfixWords")
+    val bugfixWords: Column<Double> = double("bugfixWords")
     // TODO: more metrics
 }
